feat: refactor room page and routing logic
