--- src/components/pages/room.js
@@ -1,200 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import io from "socket.io-client";
-import { Loading } from "@/components/loading";
-import { Deck } from "@/components/deck";
-import Card from "@/components/card";
-import Agreement from "@/components/agreement";
-import "react-contexify/dist/ReactContexify.css";
-import { useContextMenu } from "react-contexify";
-import { CustomMenu } from "../customMenu";
-
-let socket;
-
-export default function Room({ room, userName }) {
-  const [users, setUsers] = useState([]);
-  const [roomInfo, setRoomInfo] = useState({});
-  const [newStoryTitle, setNewStoryTitle] = useState("");
-  const [currentUserId, setCurrentUserId] = useState(null);
-  const userTileMenuId = "userTileMenuId";
-  const { show } = useContextMenu({
-    id: userTileMenuId,
-  });
-
-  useEffect(() => {
-    initSocket();
-    return () => {
-      socket.disconnect();
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const initSocket = async () => {
-    await fetch("/api/socket");
-    socket = io(undefined, {
-      path: "/api/socket_io",
-    });
-    socket.on("connect", () => {
-      setCurrentUserId(socket.id);
-      console.log("client - connected", socket.id);
-    });
-
-    socket.emit("user-joined", { room, userName });
-    socket.on("room-users", (data) => {
-      setUsers(data.users);
-      setRoomInfo(data.roomInfo);
-      console.log(data);
-    });
-  };
-
-  if (!users.length) {
-    return <Loading />;
-  }
-
-  const handleVote = (vote) => {
-    socket.emit("user-vote", { room, userName, vote });
-  };
-
-  const toggleRevealState = () => {
-    const doWeHaveMinOneVote = users.some((user) => user.vote); // To check all the users voted, change some to every
-    if (!doWeHaveMinOneVote) return;
-
-    socket.emit("room-info-update", {
-      room,
-      revealState: roomInfo.revealState === "open" ? "close" : "open",
-    });
-  };
-
-  const handleStartNewVote = () => {
-    socket.emit("room-info-update", {
-      room,
-      revealState: "close",
-      userStory: {
-        title: newStoryTitle,
-        description: null,
-      },
-      startNewVote: true,
-    });
-    document.getElementById("story-title-input-field").value = "";
-    setNewStoryTitle("");
-  };
-
-  const handleAssignAdmin = (userId) => {
-    socket.emit("room-info-update", {
-      room,
-      newAdminId: userId,
-    });
-  };
-
-  const handleUserTileClick = (userId, event) => {
-    if (roomInfo.roomAdmin !== userId && roomInfo.roomAdmin === currentUserId) {
-      show({ event, props: { userId } });
-    }
-  };
-
-  return (
-    <main
-      className="container mx-auto flex gap-5 flex-col p-2 flex-1 w-full "
-      style={{
-        height: `calc(100vh) - 70px`,
-      }}
-    >
-      <div className="flex flex-row justify-between p-2">
-        <div className="flex-1">
-          {currentUserId === roomInfo.roomAdmin && (
-            <div className="flex flex-col lg:flex-row gap-2 items-center self-start">
-              <div className="self-start gap-2 w-full lg:w-96 flex items-center px-1 py-2">
-                <input
-                  id="story-title-input-field"
-                  className="input input-bordered rounded-xl input-primary w-full"
-                  placeholder="Please enter the user story title (optional)"
-                  onChange={(event) => {
-                    setNewStoryTitle(event.target.value);
-                  }}
-                />
-              </div>
-              <div className="flex self-start gap-2">
-                <div className="cursor-pointer label">
-                  <button
-                    onClick={handleStartNewVote}
-                    className="btn btn-primary text-base-100 normal-case rounded-lg font-medium"
-                  >
-                    Start New Voting
-                  </button>
-                </div>
-                <div className="cursor-pointer label">
-                  <span className="label-text font-semibold px-2 text-lg">
-                    Reveal Votes
-                  </span>
-                  <input
-                    type="checkbox"
-                    className="toggle toggle-success toggle-lg"
-                    checked={roomInfo.revealState === "open"}
-                    onClick={toggleRevealState}
-                  />
-                </div>
-              </div>
-            </div>
-          )}
-          {roomInfo.userStory.title ? (
-            <div className="flex gap-2 items-center self-start pt-2 p-2">
-              <label className="font-medium">{`You'll estimate for :`}</label>
-              <p className="font-bold">{roomInfo.userStory.title}</p>
-            </div>
-          ) : null}
-
-          <div className="flex-1 pb-[400px]">
-            <div className="flex flex-wrap mt-2 items-center justify-center md:items-start md:justify-start">
-              {users.map((user, index) => {
-                return (
-                  <Card
-                    onClick={(event) => {
-                      handleUserTileClick(user.id, event);
-                    }}
-                    onContextMenu={(event) => {
-                      handleUserTileClick(user.id, event);
-                    }}
-                    reveal={roomInfo.revealState === "open"}
-                    isAdmin={roomInfo.roomAdmin === user.id}
-                    index={index}
-                    {...user}
-                    key={user.id}
-                  />
-                );
-              })}
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="voting-floater rounded-2xl p-3 bottom-1 drop-shadow-md bg-primary bg-opacity-10">
-        <div className="container mx-auto flex flex-col gap-5 justify-center items-center">
-          {/* {isAdmin && (
-            <div className="flex gap-2 items-center self-start lg:absolute">
-              <input
-                className="checkbox checkbox-success rounded-lg"
-                type="checkbox"
-                placeholder="Enter story title"
-              />
-              <label className="font-bold font-neutral">
-                Vote as participant
-              </label>
-            </div>
-          )} */}
-
-          {roomInfo.revealState === "open" ? (
-            <Agreement users={users} />
-          ) : (
-            <Deck onVote={handleVote} users={users} userName={userName}></Deck>
-          )}
-        </div>
-      </div>
-      <CustomMenu
-        menuId={userTileMenuId}
-        handleOnClick={(event) => {
-          handleAssignAdmin(event.props.userId);
-        }}
-      />
-    </main>
-  );
-}
