+++ src/lib/utils.js
@@ -15,8 +15,10 @@ const generateFibonacciSeries = (upto = 21) => {
   return series;
 };
 
+// should we copy the room link or landing page link
+// both will work but user will be redirected based on this value
 const getInviteRoomLink = (room) => {
-  return `${window.location.host}/?room=${room}`;
+  return `${window.location.host}/room?roomid=${room}`;
 };
 
 const writeToClipboard = async (content) => {
